+++ contracts/contracts/interfaces/IRegisterCircuitVerifier.sol
@@ -5,6 +5,7 @@ pragma solidity 0.8.28;
  * @notice Interface for verifying register circuit proofs.
  * @dev This interface defines the structure of a register circuit proof and exposes a function to verify such proofs.
  */
+
 interface IRegisterCircuitVerifier {
     /**
      * @notice Represents a register circuit proof.
@@ -15,10 +16,10 @@ interface IRegisterCircuitVerifier {
      * @param pubSignals An array of three unsigned integers representing the public signals associated with the proof.
      */
     struct RegisterCircuitProof {
-        uint[2] a;
-        uint[2][2] b;
-        uint[2] c;
-        uint[3] pubSignals;
+        uint256[2] a;
+        uint256[2][2] b;
+        uint256[2] c;
+        uint256[3] pubSignals;
     }
 
     /**
@@ -31,9 +32,9 @@ interface IRegisterCircuitVerifier {
      * @return isValid A boolean value indicating whether the provided proof is valid (true) or not (false).
      */
     function verifyProof(
-        uint[2] calldata a,
-        uint[2][2] calldata b,
-        uint[2] calldata c,
-        uint[3] calldata pubSignals
+        uint256[2] calldata a,
+        uint256[2][2] calldata b,
+        uint256[2] calldata c,
+        uint256[3] calldata pubSignals
     ) external view returns (bool isValid);
 }
